fix: retardo en inicialización de UI scene
La creación de la UI se pasa al 'update' en lugar del 'create' porque en caso contrario dicha escena trata de acceder a datos creados en 'init' sin haber dejado pasar el ciclo necesario para pasar por ahí.

Es la mejor solución que he encontrado por ahora y funciona.

Son dos ciclos de espera, el primero para la carga de datos de la escena, y el segundo para la carga de datos del jugador.

+++ src/game/player.js
@@ -37,7 +37,7 @@ export default class PlayerLogic extends Entity
         // variables de control de saltos
         this.maxJumps = 2;
         this.nextJump = this.scene.time.now;
-        this.allowedJumps = this.maxJumps;
+        this.allowedJumps = 0; // this.maxJumps;
 
         // booleanno para saber si puede ser dañado
         this.canBeDamaged = true;
@@ -60,8 +60,10 @@ export default class PlayerLogic extends Entity
         this.playerAnimation(groundCheck);
 
         // UI
-        this.scene.UI.rewriteUI(this.scene.UI.place01, 'Allowed jumps: ' + this.allowedJumps + "/" + this.maxJumps);
-        this.scene.UI.rewriteUI(this.scene.UI.place02, 'Lives: ' + this.health);
+        if (this.scene.UI.initC) {
+            this.scene.UI.rewriteUI(this.scene.UI.place01, 'Allowed jumps: ' + this.allowedJumps + "/" + this.maxJumps);
+            this.scene.UI.rewriteUI(this.scene.UI.place02, 'Lives: ' + this.health);
+        }
     }
 
     setContainer(container) { this.container = container; }
@@ -70,7 +72,7 @@ export default class PlayerLogic extends Entity
         // ground detection logic
         if (groundCheck) {
             this.allowedJumps = this.maxJumps;
-            this.scene.UI.jumps.reset(this.allowedJumps, 'jump'); // UI
+            if (this.scene.UI.initP) this.scene.UI.jumps.reset(this.allowedJumps, 'jump'); // UI
             this.velocity.y = 0;
         } else {
             this.velocity.y = this.velocity.y + this.scene.physics.config.gravity.y * dt / 1000;
@@ -136,7 +138,7 @@ export default class PlayerLogic extends Entity
         if (this.allowedJumps > 0 && this.up && this.nextJump < this.scene.time.now)
         {
             this.allowedJumps--;
-            this.scene.UI.jumps.deleteLast(); // UI
+            if (this.scene.UI.initP) this.scene.UI.jumps.deleteLast(); // UI
             // console.log('allowed jumps: ' + this.allowedJumps);
             // this.setVelocityY(-190);
             this.velocity.y = -190;
@@ -189,10 +191,10 @@ export default class PlayerLogic extends Entity
         if(this.health > this.maxHealth)
         {
             this.health = this.maxHealth
-            this.scene.UI.lives.reset(this.health, 'object'); // UI
+            if (this.scene.UI.initP) this.scene.UI.lives.reset(this.health, 'object'); // UI
         }
         else {
-            this.scene.UI.lives.addObjects(power, 'object'); // UI
+            if (this.scene.UI.initP) this.scene.UI.lives.addObjects(power, 'object'); // UI
         }
         console.log("health" + this.health)
     }
@@ -202,7 +204,7 @@ export default class PlayerLogic extends Entity
         if (this.canBeDamaged)
         {
             this.health -= power;
-            this.scene.UI.lives.deleteObjects(power); // UI
+            if (this.scene.UI.initP) this.scene.UI.lives.deleteObjects(power); // UI
             this.canBeDamaged = false;
             console.log("health: " + this.health)
